+++ Packages/com.unity.inputsystem/InputSystem/Controls/InputControlExtensions.cs
@@ -366,9 +366,10 @@ namespace UnityEngine.InputSystem
         /// <summary>
         /// Check if the given state corresponds to the default state of the control.
         /// </summary>
+        /// <param name="control">Control to check the state for in <paramref name="statePtr"/>.</param>
         /// <param name="statePtr">Pointer to a state buffer containing the <see cref="InputControl.stateBlock"/> for <paramref name="control"/>.</param>
-        /// <param name="maskPtr">If not null, only bits set to true in the buffer will be taken into account. This can be used
-        /// to mask out noise.</param>
+        /// <param name="maskPtr">If not null, only bits set to <c>false</c> (!) in the buffer will be taken into account. This can be used
+        /// to mask out noise, i.e. every bit that is set in the mask is considered to represent noise.</param>
         /// <returns>True if the control/device is in its default state.</returns>
         /// <remarks>
         /// Note that default does not equate all zeroes. Stick axes, for example, that are stored as unsigned byte