+++ Packages/com.unity.inputsystem/InputSystem/Devices/Gamepad.cs
@@ -181,6 +181,7 @@ namespace UnityEngine.InputSystem.LowLevel
         }
     }
 
+    ////NOTE: The bit positions here based on the enum value are also used in native.
     /// <summary>
     /// Enum of common gamepad buttons.
     /// </summary>
@@ -196,22 +197,22 @@ namespace UnityEngine.InputSystem.LowLevel
         /// <summary>
         /// The up button on a gamepad's dpad.
         /// </summary>
-        DpadUp,
+        DpadUp = 0,
 
         /// <summary>
         /// The down button on a gamepad's dpad.
         /// </summary>
-        DpadDown,
+        DpadDown = 1,
 
         /// <summary>
         /// The left button on a gamepad's dpad.
         /// </summary>
-        DpadLeft,
+        DpadLeft = 2,
 
         /// <summary>
         /// The right button on a gamepad's dpad.
         /// </summary>
-        DpadRight,
+        DpadRight = 3,
 
         // Face buttons. We go with a north/south/east/west naming as that
         // clearly disambiguates where we expect the respective button to be.
@@ -222,7 +223,7 @@ namespace UnityEngine.InputSystem.LowLevel
         /// <remarks>
         /// Identical to <see cref="Y"/> and <see cref="Triangle"/> which are the Xbox and PlayStation controller names for this button.
         /// </remarks>
-        North,
+        North = 4,
 
         /// <summary>
         /// The right action button on a gamepad.
@@ -230,7 +231,7 @@ namespace UnityEngine.InputSystem.LowLevel
         /// <remarks>
         /// Identical to <see cref="B"/> and <see cref="Circle"/> which are the Xbox and PlayStation controller names for this button.
         /// </remarks>
-        East,
+        East = 5,
 
         /// <summary>
         /// The lower action button on a gamepad.
@@ -238,7 +239,7 @@ namespace UnityEngine.InputSystem.LowLevel
         /// <remarks>
         /// Identical to <see cref="A"/> and <see cref="Cross"/> which are the Xbox and PlayStation controller names for this button.
         /// </remarks>
-        South,
+        South = 6,
 
         /// <summary>
         /// The left action button on a gamepad.
@@ -246,48 +247,51 @@ namespace UnityEngine.InputSystem.LowLevel
         /// <remarks>
         /// Identical to <see cref="X"/> and <see cref="Square"/> which are the Xbox and PlayStation controller names for this button.
         /// </remarks>
-        West,
+        West = 7,
 
 
         /// <summary>
         /// The button pressed by pressing down the left stick on a gamepad.
         /// </summary>
-        LeftStick,
+        LeftStick = 8,
 
         /// <summary>
         /// The button pressed by pressing down the right stick on a gamepad.
         /// </summary>
-        RightStick,
+        RightStick = 9,
 
         /// <summary>
         /// The left shoulder button on a gamepad.
         /// </summary>
-        LeftShoulder,
+        LeftShoulder = 10,
 
         /// <summary>
         /// The right shoulder button on a gamepad.
         /// </summary>
-        RightShoulder,
-
-        /// <summary>
-        /// The left trigger button on a gamepad.
-        /// </summary>
-        LeftTrigger,
-
-        /// <summary>
-        /// The right trigger button on a gamepad.
-        /// </summary>
-        RightTrigger,
+        RightShoulder = 11,
 
         /// <summary>
         /// The start button.
         /// </summary>
-        Start,
+        Start = 12,
 
         /// <summary>
         /// The select button.
         /// </summary>
-        Select,
+        Select = 13,
+
+        // For values that are not part of the buttons bitmask in GamepadState, assign large values that are outside
+        // the 32bit bit range.
+
+        /// <summary>
+        /// The left trigger button on a gamepad.
+        /// </summary>
+        LeftTrigger = 32,
+
+        /// <summary>
+        /// The right trigger button on a gamepad.
+        /// </summary>
+        RightTrigger = 33,
 
         /// <summary>
         /// The X button on an Xbox controller.